==> analysis/03_D62_Gene_Inspection.Rmd <==
---
title: "Check Markers"
author: "AGC, AY"
date: "17 3 2021"
output: html_document
---


# mRNA Expression of Celllular Markers 

```{r setup, include=FALSE}



library(DESeq2)
library(knitr)
library(tidyverse)
library(pheatmap)

opts_chunk$set(echo = TRUE, 
               fig.height = 7, 
               fig.width = 9, 
               message = TRUE,
               warning = TRUE,
               fig.align='center',
               dpi=500
)

```

```{r load_data}
home = getwd()
output= paste0(home, "/output/")

source(paste0(home,"/code/custom_functions.R"))
load(paste0(output,"/D62_dds_matrix.RData"))

```


## Full Marker list 

```{r plotGOIs all markers, fig.height=15, fig.width=8}

SampleInfo = as.data.frame(colData(ddsMat))
genedata=rowData(ddsMat)

SampleInfo = SampleInfo %>% dplyr::arrange(desc(gRNA), desc(DIFF), desc(RAPA))

# image(SampleInfo %>% select( CellLine, gRNA, DIFF, RAPA) %>% mutate_all(as.numeric) %>% as.matrix)

GOIsEntrez=c(DEPDC5="9681",
             C_MYC="4609",
             P53="7157",
             PIK3CA="5290",
             REDD1="54541",
             BNIP3="664",
             HRAS="3265",
             KRAS="3845",
             NRAS="4893",
             AKT="207",
             TSC1="7248",
             TSC2="7249",
             MTOR="2475",
             RPS6="6194",
             RPTOR="57521",
             RAG_A="10670",
             RAG_B="10325",
             RAG_C="64121",
             RAG_D="58528",
             DEPTOR="64798",
             RHEB="6009",
             MAPK="5594",
             NPRL2="10641",
             NPRL3="8131",
             CASTOR1="652968",
             CASTOR2="729438",
             CASTOR3="352954",
             SESTRIN_1="27244",
             SESTRIN_2="83667",
             SESTRIN_3="143686",
             NCAM1="4684",
             EGF="1950",
             EGFR="1956",
             FGFR1="2260",
             FGFR2="2263",
             FGFR3="2261",
             FGFR4="2264",
             FGF1="2246",
             FGF2="2247",
             TRKC="4916",
             NT3="4908",
             TRKB="4915",
             NGFR="4804",
             PRKACA="5566",
             PRKACB="5567",
             PRKAR1A="5573",
             PRKAR2B="5577",
             PRKCA="5578",
             OCT4="5460",
             NANOG="79923",
             MSI1="4440",
             SOX1="6656",
             TBR1=	"10716",
             TBR2=	"8320",
             MASH1_ASCL1="429",
             HES5="388585",
             PAX6="5080",
             SOX2="6657",
             NESTIN="10763",
             GFAP="2670",
             S100B="6285",
             AQP4="361",
             SOX9="6662",
             CX3CR1="1524",
             OLIG2="10215",
             GRIA2="2891",
             GRIN1	="2902",
             GRM1="2911",
             GRIN2A="2903",
             VGLUT1="57030",
             GLAST_SLC1A3="6507",
             GLUL="2752",
             GAD1="2571",	
             GAT1=	"6529",	
             ENO2=	"2026",	
             MAPT=	"4137",	
             DCX="1641",
             MAP2="4133",
             TUBB3="10381",
             NEUN="146713",
             PSD95="1742",
             HOMER1="9456",
             ACHE=	"43",
             HPRT1=	"3251",
             NET1=	"10276",
             POU3F2=	"5454",
             TH="7054",
             SYNAPTOPHYSIN="6855",
             GAP_43="2596",
             BASSOON="8927",
             STXBP1="6812",
             STX1A="6804",
             SNCA="6622")


geneheatmap(GOIsEntrez,ddsMat)

```

## Set of 15 Cellular Markers

```{r plotGOIs 15 markers, fig.width=9, fig.height=5}

GOIsEntrez=c(
  DEPDC5="9681",
  OCT4="5460",
  NANOG="79923",
  SOX2="6657",
  NESTIN="10763",
  PAX6="5080",
  GFAP="2670",
  AQP4="361",
  ASCL1="429",
  Olig2="10215",
  TMEM119="338773",
  TUBB3="10381",
  MAP2="4133",
  SYP="6855",
  SLC17A7="57030",
  GRIA2="2891",
  GAD1="2571",
  GABBR2="9568",
  SERT1="6532",
  HTR2B=" 3357",
  CHAT="1103",
  CHRNA7="1139",
  TH="7054",
  DRD2="1813")

GOIplot = geneheatmap(GOIsEntrez,ddsMat, show_rownames = T)

pdf("./docs/Publication_Supplfigure4A.pdf", width = 9, height = 5)
GOIplot
dev.off()

ggsave("./docs/Publication_Supplfigure4A.svg", GOIplot, width =9, height = 5)


```


## Cellular Markers Proliferation Subset

```{r plotGOIs 15 markers Proliferation, fig.width=8}

geneheatmap(GOIsEntrez,ddsMat[,ddsMat$DIFF=="noDIFF"])

```


## Cellular Markers Differentation Subset

```{r plotGOIs 15 markers Differentation, fig.width=8}

geneheatmap(GOIsEntrez,ddsMat[,ddsMat$DIFF=="DIFF"])

```
