==> analysis/05_D62_Statistical_analysis.Rmd <==
---
title: "Statistical analysis"
author: "AGC, AY"
date: "23 3 2021"
output: html_document
---

```{r}
home = getwd()

knitr::opts_chunk$set(fig.height=6, fig.width=8)


```

# Statistical analysis 

```{r setup, include=FALSE, echo=F}
reanalyze=F
output= paste0(home, "/output/")

source(paste0(home,"/code/custom_functions.R"))

library(DESeq2)
library(knitr)
library(tidyverse)
library(pheatmap)
library(WGCNA)
library(flashClust)
library(pheatmap)
library(limma)
library(openxlsx)
library(eulerr)

opts_chunk$set(echo = TRUE, 
               fig.height = 7, 
               fig.width = 9, 
               message = TRUE,
               warning = TRUE,
               fig.align='center',
               dpi=500)

load(paste0(output,"/D62_dds_matrix.RData"))
SampleInfo = as.data.frame(colData(ddsMat))
log_2cpm=log2(counts(ddsMat, normalize=T)+1)

```



# Which genes are differentially regualted upon KO of DEPDC5 

## Modelling the KO in all different combinations

```{r loops_for_comparison, message=F}

if(reanalyze | !file.exists(paste0(home,"/output/D62_ResTabs_KO.RData"))){
  
  
  ## comparisons against noRAPA NTC
  Rapamycin=c("noRAPA", "RAPA")
  Differentiation=c("noDIFF", "DIFF")
  Type_sgRNA<-c("sg2.1","sg2.2")
  target="KO"
  
  r = Rapamycin[1]
  d = Differentiation[1]
  Tp = Type_sgRNA[[1]]
  
  
  # no random effects included 
  for(r in Rapamycin){
    Rapafilter = SampleInfo$RAPA %in% r
    for(d in Differentiation){
      Difffilter = SampleInfo$DIFF %in% d
      for(Tp in Type_sgRNA){
        sgRNAfilter = SampleInfo$gRNA %in% Tp
        vs_label=paste0(c("sgNTC", Tp), sep="", collapse="_")
        Set = rownames(SampleInfo)[Rapafilter&Difffilter&
                                     sgRNAfilter]
        Set = c(Set, rownames(SampleInfo)[SampleInfo$RAPA == "noRAPA" & Difffilter&
                                            SampleInfo$gRNA == "sgNTC"])
        
        lab = paste("restab", "D62_NTCnoRAPA", vs_label, d,r, sep="_")
        print(lab)
        assign(lab, comparison(ddsMat, Set, target, randomeffect = NULL))
      }
    }
  }
  
  
  # calculate all combinations against same NTC (treated)
  
  Type_sgRNA<-list(c("sgNTC","sg2.1"),c("sgNTC","sg2.2"))
  Cl<-"D62"
  
  # no random effects included 
  for(r in Rapamycin){
    Rapafilter = SampleInfo$RAPA %in% r
    for(d in Differentiation){
      Difffilter = SampleInfo$DIFF %in% d
      for(Tp in Type_sgRNA){
        sgRNAfilter = SampleInfo$gRNA %in% Tp
        vs_label=paste0(Tp, sep="", collapse="_")
        
        
        Set = rownames(SampleInfo)[Rapafilter&Difffilter&
                                     sgRNAfilter]
        lab = paste("restab", "D62_NTCwRAPA",vs_label, d,r, sep="_")
        print(lab)
        assign(lab, 
               comparison(ddsMat, samples = Set, 
                          target =target,randomeffect = c()))
      }
    }
  }
  
  comparisons= apropos("restab")
  save(list = comparisons, file = paste0(home,"/output/D62_ResTabs_KO.RData"))
} else 
  load(file = paste0(home,"/output/D62_ResTabs_KO.RData"))

```


```{r predef_for_analaysis}
samesign <- function(x) {abs(sum(sign(x)))==length(x)}
mypval=0.05
colors <- rev(colorRampPalette(brewer.pal(9, "Spectral"))(255))


genelists=lapply(list.files(paste0(home,"/data/genelists_to_test"), full.names = T), 
                 function(x){read.table(x, header=T)})

names(genelists) = gsub(".txt", "", list.files(paste0(home,"/data/genelists_to_test")))



getlinmodoutput = function(targetline="D62", 
                           targetdiff, 
                           targetrapa, 
                           refset="NTCnoRAPA"){
  
  
  if(refset=="NTCnoRAPA"){
    samplesincl=SampleInfo$DIFF==targetdiff &
      SampleInfo$RAPA==targetrapa &  
      SampleInfo$KO == "KO"
    samplesincl = samplesincl | (SampleInfo$DIFF==targetdiff &
                                   SampleInfo$RAPA=="noRAPA" &  
                                   SampleInfo$KO == "WT")
    
  } else {
    samplesincl = SampleInfo$DIFF==targetdiff & 
      SampleInfo$RAPA==targetrapa 
  }  
  
  
  pvalrep=  get(paste0("restab_", 
                       targetline,"_",refset, "_sgNTC_sg2.1_", 
                       targetdiff, "_" , 
                       targetrapa))$padj<mypval &
    get(paste0("restab_", 
               targetline, "_",refset,"_sgNTC_sg2.2_", 
               targetdiff, "_" , 
               targetrapa))$padj<mypval
  
  betarep = apply(cbind(get(paste0("restab_", 
                                   targetline,"_",refset,  "_sgNTC_sg2.1_", 
                                   targetdiff, "_" , 
                                   targetrapa))$log2FoldChange,
                        get(paste0("restab_", 
                                   targetline,"_",refset,  "_sgNTC_sg2.2_", 
                                   targetdiff, "_" , 
                                   targetrapa))$log2FoldChange), 1,
                  samesign)
  
  idx=which(betarep & pvalrep)
  hits=geneids[idx]
  
  restab=data.frame(
    entrezgene = geneids[idx], 
    genname=rowData(ddsMat)[hits,"hgnc"],
    log2FC_2.1 = get(paste0("restab_", 
                            targetline,"_",refset,  "_sgNTC_sg2.1_", 
                            targetdiff, "_" , 
                            targetrapa))$log2FoldChange[idx], 
    fdr_2.1 = get(paste0("restab_", 
                         targetline,"_",refset,  "_sgNTC_sg2.1_", 
                         targetdiff, "_" , 
                         targetrapa))$padj[idx], 
    log2FC_2.2 = get(paste0("restab_", 
                            targetline,"_",refset,  "_sgNTC_sg2.2_", 
                            targetdiff, "_" , 
                            targetrapa))$log2FoldChange[idx], 
    fdr_2.2 = get(paste0("restab_", 
                         targetline,"_",refset,  "_sgNTC_sg2.2_", 
                         targetdiff, "_" , 
                         targetrapa))$padj[idx])
  
  print(restab)
  write.xlsx(restab, file=paste0(output, "/Restab_", 
                                 targetline,"_",refset, "_", 
                                 targetdiff, "_",
                                 targetrapa, ".xlsx"))
  
  SamplesSet=SampleInfo[samplesincl,]
  
  plotmatrix = log_2cpm[hits,rownames(SamplesSet)]
  
  rownames(SamplesSet)=SamplesSet$label_rep
  
  colnames(plotmatrix)=SamplesSet$label_rep
  rownames(plotmatrix)=rowData(ddsMat)[hits,"hgnc"]     
  
  
  gRNAcol = Dark8[c(1:nlevels(SampleInfo$gRNA))+nlevels(SampleInfo$CellLine)]
  names(gRNAcol) = levels(SampleInfo$gRNA)
  gRNAcol = gRNAcol[as.character(unique(SamplesSet$gRNA))]
  
  diffcol = brewer.pal(3,"Set1")[1:nlevels(SampleInfo$DIFF)]
  names(diffcol) = levels(SampleInfo$DIFF)
  diffcol = diffcol[as.character(unique(SamplesSet$DIFF))]
  
  rapacol = brewer.pal(3,"Set2")[1:nlevels(SampleInfo$RAPA)]
  names(rapacol) = levels(SampleInfo$RAPA)
  rapacol = rapacol[as.character(unique(SamplesSet$RAPA))]
  
  ann_colors = list(
    DIFF = diffcol, 
    RAPA = rapacol,
    gRNA = gRNAcol)
  
  collabels = SamplesSet[,c("DIFF","RAPA", "gRNA")] %>%  
    mutate_all(as.character) %>% as.data.frame()
  
  rownames(collabels)=SamplesSet$label_rep
  
  idx=order(SamplesSet$gRNA)
  
  
  pheatmap(plotmatrix[,idx],
           border_color = NA, 
           annotation_col = collabels[idx,],
           show_rownames = F, show_colnames = F,
           annotation_colors = ann_colors,
           clustering_method = "ward.D2",
           cluster_cols = F,
           col = colors, 
           scale = "row",
           
           main = paste("Normalized log2 counts", targetline, targetdiff, targetrapa))
  
  gene_univers = rownames(ddsMat)
  
  gostres = getGOresults(hits, gene_univers, evcodes = T)
  
  toptab = gostres$result
  
  
  write.xlsx(toptab, file = paste0(output, "/GOres_",targetline,targetdiff,targetrapa,".xlsx"), sheetName = "GO_enrichment")
  
  idx = grepl("GO|KEGG", toptab$source)
  titlename = paste(targetline,targetdiff,targetrapa)
  if(!any(idx)){
    p = ggplot() + annotate("text", x = 4, y = 25, size=4, 
                            label = "no significant GO term") + 
      ggtitle(titlename)+theme_void()+
      theme(plot.title = element_text(hjust = 0.5))
  } else {
    p=GOplot(toptab[idx, ], 10, Title = titlename)
  }
  print(p)
  
  ## enrichment test for genelist 
  Resall=data.frame()
  
  for (gl in names(genelists)){
    tmp=table(signif=geneids %in% hits, targetgene=geneids %in% genelists[[gl]]$ENTREZ_ID)
    resfish=fisher.test(tmp)
    res = c(resfish$estimate, unlist(resfish$conf.int), resfish$p.value)
    Resall = rbind(Resall, res)
  }
  colnames(Resall)=c("OR", "CI95L", "CI95U", "P")
  rownames(Resall)=names(genelists)
  Resall$Beta = log(Resall$OR)
  Resall$SE = (log(Resall$OR)-log(Resall$CI95L))/1.96
  Resall$Padj=p.adjust(Resall$P, method = "bonferroni")
  
  print(Resall)
  
  multiORplot(Resall, Pval = "P", Padj = "Padj", beta="Beta",SE = "SE", 
              pheno=paste(targetline, targetdiff, targetrapa), xlims=c(-4.5,4.5))
}


geneids=rownames(restab_D62_NTCnoRAPA_sgNTC_sg2.1_DIFF_noRAPA)

```

## KO effect in noDIFF noRAPA
### D62

```{r noDIFF_noRAPA_D62}

getlinmodoutput(targetdiff = "noDIFF",targetrapa = "noRAPA")

```

## KO effect in DIFF noRAPA

### D62

```{r DIFF_noRAPA_D62}

getlinmodoutput(targetdiff = "DIFF",targetrapa = "noRAPA")

```

## KO effect in noDIFF RAPA
### D62

```{r noDIFF_RAPA_D62}

getlinmodoutput(targetdiff = "noDIFF",targetrapa = "RAPA")
getlinmodoutput(targetdiff = "noDIFF",targetrapa = "RAPA", refset = "NTCwRAPA" )

```

## KO effect in DIFF RAPA

### D62

```{r DIFF_RAPA_D62}

getlinmodoutput(targetdiff = "DIFF",targetrapa = "RAPA")
getlinmodoutput(targetdiff = "DIFF",targetrapa = "RAPA", refset = "NTCwRAPA" )

```


```{r custom table for paper}
# with noRAPA NTCF as reference

noDIFFresults=data.frame(ENTREZ_ID=rownames(restab_D62_NTCnoRAPA_sgNTC_sg2.1_noDIFF_noRAPA),
                         SYMBOL=rowData(ddsMat)$hgnc,
                         REPLICATED = restab_D62_NTCnoRAPA_sgNTC_sg2.1_noDIFF_noRAPA$log2FoldChange*
                           restab_D62_NTCnoRAPA_sgNTC_sg2.2_noDIFF_noRAPA$log2FoldChange>0 & 
                           restab_D62_NTCnoRAPA_sgNTC_sg2.1_noDIFF_noRAPA$padj<=.05  &
                           restab_D62_NTCnoRAPA_sgNTC_sg2.2_noDIFF_noRAPA$padj<=.05,
                         sg2.1_noDIFF_noRAPA_FC=restab_D62_NTCnoRAPA_sgNTC_sg2.1_noDIFF_noRAPA$log2FoldChange,
                         sg2.1_noDIFF_noRAPA_adjP=restab_D62_NTCnoRAPA_sgNTC_sg2.1_noDIFF_noRAPA$padj,
                         sg2.2_noDIFF_noRAPA_FC=restab_D62_NTCnoRAPA_sgNTC_sg2.2_noDIFF_noRAPA$log2FoldChange,
                         sg2.2_noDIFF_noRAPA_adjP=restab_D62_NTCnoRAPA_sgNTC_sg2.2_noDIFF_noRAPA$padj,
                         sgAVG_noDIFF_noRAPA_FC = rowMeans(cbind(restab_D62_NTCnoRAPA_sgNTC_sg2.1_noDIFF_noRAPA$log2FoldChange,                                                                 
                                                                 restab_D62_NTCnoRAPA_sgNTC_sg2.2_noDIFF_noRAPA$log2FoldChange)),
                         sgMax_noDIFF_noRAPA_adjP = rowMaxs(cbind(restab_D62_NTCnoRAPA_sgNTC_sg2.1_noDIFF_noRAPA$padj,                                                                 
                                                                  restab_D62_NTCnoRAPA_sgNTC_sg2.2_noDIFF_noRAPA$padj)),
                         
                         sg2.1_noDIFF_RAPA_FC=restab_D62_NTCnoRAPA_sgNTC_sg2.1_noDIFF_RAPA$log2FoldChange,
                         sg2.1_noDIFF_RAPA_adjP=restab_D62_NTCnoRAPA_sgNTC_sg2.1_noDIFF_RAPA$padj,
                         sg2.2_noDIFF_RAPA_FC=restab_D62_NTCnoRAPA_sgNTC_sg2.2_noDIFF_RAPA$log2FoldChange,
                         sg2.2_noDIFF_RAPA_adjP=restab_D62_NTCnoRAPA_sgNTC_sg2.2_noDIFF_RAPA$padj, 
                         sgAVG_noDIFF_RAPA_FC = rowMeans(cbind(restab_D62_NTCnoRAPA_sgNTC_sg2.1_noDIFF_RAPA$log2FoldChange,                                                                 
                                                               restab_D62_NTCnoRAPA_sgNTC_sg2.2_noDIFF_RAPA$log2FoldChange)),
                         sgMax_noDIFF_RAPA_adjP = rowMaxs(cbind(restab_D62_NTCnoRAPA_sgNTC_sg2.1_noDIFF_RAPA$padj,                                                                 
                                                               restab_D62_NTCnoRAPA_sgNTC_sg2.2_noDIFF_RAPA$padj)),
                         
                         sg2.1_noDIFF_RAPA_rapantc_FC=restab_D62_NTCwRAPA_sgNTC_sg2.1_noDIFF_RAPA$log2FoldChange,
                         sg2.1_noDIFF_RAPA_rapantc_adjP=restab_D62_NTCwRAPA_sgNTC_sg2.1_noDIFF_RAPA$padj,
                         sg2.2_noDIFF_RAPA_rapantc_FC=restab_D62_NTCwRAPA_sgNTC_sg2.2_noDIFF_RAPA$log2FoldChange,
                         sg2.2_noDIFF_RAPA_rapantc_adjP=restab_D62_NTCwRAPA_sgNTC_sg2.2_noDIFF_RAPA$padj, 
                         sgAVG_noDIFF_RAPA_rapantc_FC = rowMeans(cbind(restab_D62_NTCwRAPA_sgNTC_sg2.1_noDIFF_RAPA$log2FoldChange,                                                         
                                                                       restab_D62_NTCwRAPA_sgNTC_sg2.2_noDIFF_RAPA$log2FoldChange)), 
                         sgMax_noDIFF_RAPA_rapantc_adjP = rowMaxs(cbind(restab_D62_NTCwRAPA_sgNTC_sg2.1_noDIFF_RAPA$padj,                                                         
                                                                        restab_D62_NTCwRAPA_sgNTC_sg2.2_noDIFF_RAPA$padj)))
for(i in names(genelists)){
  noDIFFresults[,i] <- noDIFFresults$ENTREZ_ID %in% genelists[[i]]$ENTREZ_ID
}

write.xlsx(noDIFFresults, file=paste0(output, "/noDIFF_refnoRAPA_allgenes.xlsx"))

lm(noDIFFresults$sgAVG_noDIFF_noRAPA_FC~noDIFFresults$sgAVG_noDIFF_RAPA_FC) %>% summary()

DIFFresults=data.frame(ENTREZ_ID=rownames(restab_D62_NTCnoRAPA_sgNTC_sg2.1_DIFF_noRAPA),
                       SYMBOL=rowData(ddsMat)$hgnc,
                       REPLICATED = restab_D62_NTCnoRAPA_sgNTC_sg2.1_DIFF_noRAPA$log2FoldChange*
                         restab_D62_NTCnoRAPA_sgNTC_sg2.2_DIFF_noRAPA$log2FoldChange>0 & 
                         restab_D62_NTCnoRAPA_sgNTC_sg2.1_DIFF_noRAPA$padj<=.05  &
                         restab_D62_NTCnoRAPA_sgNTC_sg2.2_DIFF_noRAPA$padj<=.05,
                       sg2.1_DIFF_noRAPA_FC=restab_D62_NTCnoRAPA_sgNTC_sg2.1_DIFF_noRAPA$log2FoldChange,
                       sg2.1_DIFF_noRAPA_adjP=restab_D62_NTCnoRAPA_sgNTC_sg2.1_DIFF_noRAPA$padj,
                       sg2.2_DIFF_noRAPA_FC=restab_D62_NTCnoRAPA_sgNTC_sg2.2_DIFF_noRAPA$log2FoldChange,
                       sg2.2_DIFF_noRAPA_adjP=restab_D62_NTCnoRAPA_sgNTC_sg2.2_DIFF_noRAPA$padj,
                       sgAVG_DIFF_noRAPA_FC = rowMeans(cbind(restab_D62_NTCnoRAPA_sgNTC_sg2.1_DIFF_noRAPA$log2FoldChange,                                                                 restab_D62_NTCnoRAPA_sgNTC_sg2.2_DIFF_noRAPA$log2FoldChange)),
                       sgMAX_DIFF_noRAPA_adjP = rowMaxs(cbind(restab_D62_NTCnoRAPA_sgNTC_sg2.1_DIFF_noRAPA$padj,                                                                 restab_D62_NTCnoRAPA_sgNTC_sg2.2_DIFF_noRAPA$padj)),
                       sg2.1_DIFF_RAPA_FC=restab_D62_NTCnoRAPA_sgNTC_sg2.1_DIFF_RAPA$log2FoldChange,
                       sg2.1_DIFF_RAPA_adjP=restab_D62_NTCnoRAPA_sgNTC_sg2.1_DIFF_RAPA$padj,
                       sg2.2_DIFF_RAPA_FC=restab_D62_NTCnoRAPA_sgNTC_sg2.2_DIFF_RAPA$log2FoldChange,
                       sg2.2_DIFF_RAPA_adjP=restab_D62_NTCnoRAPA_sgNTC_sg2.2_DIFF_RAPA$padj, 
                       sgAVG_DIFF_RAPA_FC = rowMeans(cbind(restab_D62_NTCnoRAPA_sgNTC_sg2.1_DIFF_RAPA$log2FoldChange,                                                                 restab_D62_NTCnoRAPA_sgNTC_sg2.2_DIFF_RAPA$log2FoldChange)),
                       sgMAX_DIFF_RAPA_adjP = rowMaxs(cbind(restab_D62_NTCnoRAPA_sgNTC_sg2.1_DIFF_RAPA$padj,                                                                 restab_D62_NTCnoRAPA_sgNTC_sg2.2_DIFF_RAPA$padj)),
                       
                       sg2.1_DIFF_RAPA_rapantc_FC=restab_D62_NTCwRAPA_sgNTC_sg2.1_DIFF_RAPA$log2FoldChange,
                       sg2.1_DIFF_RAPA_rapantc_adjP=restab_D62_NTCwRAPA_sgNTC_sg2.1_DIFF_RAPA$padj,
                       sg2.2_DIFF_RAPA_rapantc_FC=restab_D62_NTCwRAPA_sgNTC_sg2.2_DIFF_RAPA$log2FoldChange,
                       sg2.2_DIFF_RAPA_rapantc_adjP=restab_D62_NTCwRAPA_sgNTC_sg2.2_DIFF_RAPA$padj, 
                       sgAVG_DIFF_RAPA_rapantc_FC = rowMeans(cbind(restab_D62_NTCwRAPA_sgNTC_sg2.1_DIFF_RAPA$log2FoldChange,                                                         restab_D62_NTCwRAPA_sgNTC_sg2.2_DIFF_RAPA$log2FoldChange)), 
                       sgMAX_DIFF_RAPA_rapantc_adjP = rowMaxs(cbind(restab_D62_NTCwRAPA_sgNTC_sg2.1_DIFF_RAPA$padj,                                                         restab_D62_NTCwRAPA_sgNTC_sg2.2_DIFF_RAPA$padj)))


for(i in names(genelists)){
  DIFFresults[,i] <- DIFFresults$ENTREZ_ID %in% genelists[[i]]$ENTREZ_ID
}

metadata <- mcols(ddsMat) %>% as.data.frame()

for(i in names(genelists)){
  metadata[,i] <- rownames(metadata) %in% genelists[[i]]$ENTREZ_ID
}

mcols(ddsMat) <- metadata

write.xlsx(DIFFresults, file=paste0(output, "/DIFF_refnoRAPA_allgenes.xlsx"))

# check if the same genes are differentially reguatled 
lm(DIFFresults$sgAVG_DIFF_noRAPA_FC~DIFFresults$sgAVG_DIFF_RAPA_FC) %>% summary()

```


```{r noDIFF custom table for paper , fig.width=8, fig.height=6}
noDIFFhits=(noDIFFresults$sg2.1_noDIFF_noRAPA_adjP<=0.05 & 
              noDIFFresults$sg2.2_noDIFF_noRAPA_adjP<=0.05 &
              sign(noDIFFresults$sg2.1_noDIFF_noRAPA_FC)==sign(noDIFFresults$sg2.2_noDIFF_noRAPA_FC))

sum(noDIFFhits, na.rm = T)
sum(noDIFFhits & noDIFFresults$sg2.1_noDIFF_noRAPA_FC>0, na.rm = T)
sum(noDIFFhits & noDIFFresults$sg2.1_noDIFF_noRAPA_FC<0, na.rm = T)


ggplot(data = noDIFFresults[,], aes(sg2.1_noDIFF_noRAPA_FC, sg2.2_noDIFF_noRAPA_FC, col=noDIFFhits)) + 
  geom_point(pch=16, alpha=0.5) + 
  geom_abline(slope=1, intercept = 0) + ggtitle("DIFF correlation between KOs")



## rescued Genes 
noDIFF_RAPA_nosig = (noDIFFresults$sg2.1_noDIFF_RAPA_adjP>0.05 & noDIFFresults$sg2.2_noDIFF_RAPA_adjP>0.05)

# just necessary to check one KO as significant DEX is defoned that both have to have the same direction 

noDIFF_RAPA_chngsFC = (sign(noDIFFresults$sg2.1_noDIFF_noRAPA_FC) != sign(noDIFFresults$sg2.1_noDIFF_RAPA_FC)) &
  (sign(noDIFFresults$sg2.1_noDIFF_noRAPA_FC) != sign(noDIFFresults$sg2.2_noDIFF_RAPA_FC))


noDIFFhits_ENTREZ = na.omit(noDIFFresults$ENTREZ_ID[noDIFFhits])
noDIFFRAPA_RESC_ENTREZ = na.omit(noDIFFresults$ENTREZ_ID[ (noDIFFhits & noDIFF_RAPA_nosig)| (noDIFFhits & noDIFF_RAPA_chngsFC)])
noDIFFRAPA_noRESC_ENTREZ = noDIFFhits_ENTREZ[! noDIFFhits_ENTREZ %in% noDIFFRAPA_RESC_ENTREZ]


GOIs =list(noDIFFRAPA_RESC=noDIFFRAPA_RESC_ENTREZ,
           noDIFFRAPA_noRESC=noDIFFRAPA_noRESC_ENTREZ)

Rescue_Genelist = list(PROLIF=GOIs)

for (n in names(GOIs)){
  Resall=data.frame()
  for (gl in names(genelists)){
    
    tmp=table(signif=geneids %in% GOIs[[n]], targetgene=geneids %in% genelists[[gl]]$ENTREZ_ID)
    resfish=fisher.test(tmp)
    res = c(resfish$estimate, unlist(resfish$conf.int), resfish$p.value)
    Resall = rbind(Resall, res)
  }
  colnames(Resall)=c("OR", "CI95L", "CI95U", "P")
  rownames(Resall)=names(genelists)
  Resall$Beta = log(Resall$OR)
  Resall$SE = (log(Resall$OR)-log(Resall$CI95L))/1.96
  Resall$Padj=p.adjust(Resall$P, method = "bonferroni")
  
  print(Resall)
  
  multiORplot(Resall, Pval = "P", Padj = "Padj", beta="Beta",SE = "SE", 
              pheno=n, xlims=c(-4.5,4.5))
  
}


gene_univers = rownames(ddsMat)

gostres = getGOresults(GOIs, gene_univers)

toptab = gostres$result
write.xlsx(toptab, file = paste0(output, "/GOres_Rescue_noDIFF.xlsx"), sheetName = "GO_enrichment")


fit<- euler(list(PROLIF_DEX=noDIFFhits_ENTREZ,
                 RAPA_RESC=noDIFFRAPA_RESC_ENTREZ), shape = "ellipse")
plot(fit,  quantities = TRUE, main="PROLIF rescue")


```




```{r DIFF custom table for paper, fig.width=8, fig.height=6}
### DIFFERERNTIATION


DIFFhits=(DIFFresults$sg2.1_DIFF_noRAPA_adjP<=0.05 & 
            DIFFresults$sg2.2_DIFF_noRAPA_adjP<=0.05 &
            sign(DIFFresults$sg2.1_DIFF_noRAPA_FC)==sign(DIFFresults$sg2.2_DIFF_noRAPA_FC))



sum(DIFFhits, na.rm = T)
sum(DIFFhits & DIFFresults$sg2.1_DIFF_noRAPA_FC>0, na.rm = T)
sum(DIFFhits & DIFFresults$sg2.1_DIFF_noRAPA_FC<0, na.rm = T)


ggplot(data = DIFFresults[,], aes(sg2.1_DIFF_noRAPA_FC, sg2.2_DIFF_noRAPA_FC, col=DIFFhits)) + 
  geom_point(pch=16, alpha=0.5) + 
  geom_abline(slope=1, intercept = 0) + 
  theme_bw()



## rescued Genes 
DIFF_RAPA_nosig = (DIFFresults$sg2.1_DIFF_RAPA_adjP>0.05 & DIFFresults$sg2.2_DIFF_RAPA_adjP>0.05)

# just necessary to check one KO as significant DEX is defoned that both have to have the same direction 

DIFF_RAPA_chngsFC = (sign(DIFFresults$sg2.1_DIFF_noRAPA_FC) != sign(DIFFresults$sg2.1_DIFF_RAPA_FC)) &
  (sign(DIFFresults$sg2.1_DIFF_noRAPA_FC) != sign(DIFFresults$sg2.2_DIFF_RAPA_FC))


DIFFhits_ENTREZ = na.omit(DIFFresults$ENTREZ_ID[DIFFhits])
DIFFRAPA_RESC_ENTREZ = na.omit(DIFFresults$ENTREZ_ID[ (DIFFhits & DIFF_RAPA_nosig)| (DIFFhits & DIFF_RAPA_chngsFC)])
DIFFRAPA_noRESC_ENTREZ = DIFFhits_ENTREZ[! DIFFhits_ENTREZ %in% DIFFRAPA_RESC_ENTREZ]


GOIs =list(DIFFRAPA_RESC=DIFFRAPA_RESC_ENTREZ,
           DIFFRAPA_noRESC=DIFFRAPA_noRESC_ENTREZ)

Rescue_Genelist[["DIFF"]] = GOIs


for (n in names(GOIs)){
  Resall=data.frame()
  for (gl in names(genelists)){
    
    tmp=table(signif=geneids %in% GOIs[[n]], targetgene=geneids %in% genelists[[gl]]$ENTREZ_ID)
    resfish=fisher.test(tmp)
    res = c(resfish$estimate, unlist(resfish$conf.int), resfish$p.value)
    Resall = rbind(Resall, res)
  }
  colnames(Resall)=c("OR", "CI95L", "CI95U", "P")
  rownames(Resall)=names(genelists)
  Resall$Beta = log(Resall$OR)
  Resall$SE = (log(Resall$OR)-log(Resall$CI95L))/1.96
  Resall$Padj=p.adjust(Resall$P, method = "bonferroni")
  
  print(Resall)
  
  multiORplot(Resall, Pval = "P", Padj = "Padj", beta="Beta",SE = "SE", 
              pheno=n, xlims=c(-4.5,4.5))
  
}


gene_univers = rownames(ddsMat)

gostres = getGOresults(GOIs, gene_univers)

toptab = gostres$result
write.xlsx(toptab, file = paste0(output, "/GOres_Rescue_DIFF.xlsx"), sheetName = "GO_enrichment")


fit<- euler(list(DIFF_DEX=DIFFhits_ENTREZ,
                 RAPA_RESC=DIFFRAPA_RESC_ENTREZ), shape = "ellipse")
plot(fit,  quantities = TRUE, main="DIFF rescue")



```


```{r full overlap, fig.width=8, fig.height=6}


allhits = list(DIFF_DEX=DIFFhits_ENTREZ,
               DIFF_RAPA_RESC=DIFFRAPA_RESC_ENTREZ, 
               PROLIF_DEX=noDIFFhits_ENTREZ,
               PROLIF_RAPA_RESC=noDIFFRAPA_RESC_ENTREZ
)

fit2 <- euler(allhits, shape = "ellipse")
plot(fit2,  quantities = TRUE, main="Overlap of gene hits")

```
```{r}
save(ddsMat, file="./output/D62_dds_matrix.RData")
save(Rescue_Genelist, file = "./output/RescueGenelists.RData")

```
