==> analysis/06_D62_FiguresPublication.Rmd <==
---
title: "Figures"
author: "Chioccheti AG"
date: "2022-10-12"
output: html_document
---
```{r}
home = getwd()

knitr::opts_chunk$set(fig.height=6, fig.width=8, echo = TRUE)


```

```{r setup, include=FALSE}
library(ggpubr)
library(ggplot2)
library(ggplotify)
library(DESeq2)
library(openxlsx)
```


```{r load}
output = paste0(home, "/output/")

source(paste0(home,"/code/custom_functions.R"))

#load and prepare data
load(paste0(output,"/D62_dds_matrix.RData"))

genelists=lapply(list.files(paste0(home,"/data/genelists_to_test"), full.names = T), 
                 function(x){read.table(x, header=T)})
coldata=as.data.frame(colData(ddsMat))
metadata = mcols(ddsMat) %>% as.data.frame()
coldata$DIFF = factor(coldata$DIFF, levels = c("DIFF", "noDIFF"), labels = c("DIFF", "PROLIF"))
coldata$DIFF = relevel(coldata$DIFF, ref = "PROLIF")

logreads = log2(counts(ddsMat, normalized=T)+1)


GOresWGCNA = read.xlsx("./output/D62_GOresWGCNA.xlsx", sheet = "GO_enrichment")

prolif_dex_results = read.xlsx("./output/noDiff_refnoRAPA_allgenes.xlsx")
prolifGO_results= read.xlsx("./output/GOres_D62noDIFFnoRAPA.xlsx")
diff_dex_results = read.xlsx("./output/Diff_refnoRAPA_allgenes.xlsx")
diffGO_results= read.xlsx("./output/GOres_D62DIFFnoRAPA.xlsx")


Rescue_PROLIF_GO_results= read.xlsx("./output/GOres_Rescue_noDIFF.xlsx")
Rescue_DIFF_GO_results= read.xlsx("./output/GOres_Rescue_DIFF.xlsx")

load(file = "./output/RescueGenelists.RData")

names(genelists) = gsub(".txt", "", list.files(paste0(home,"/data/genelists_to_test")))

```


```{r}
recolor = function(p){ 
  g = ggplot_gtable(ggplot_build(p))
  stripr <- which(grepl('strip-r', g$layout$name))
  stript <- which(grepl('strip-t', g$layout$name))
  fillsr <- brewer.pal(3,"Set1")[1:2]
  fillst <- brewer.pal(3,"Set2")[1:2]
  k <- 1
  for (i in stripr) {
    j <- which(grepl('rect', g$grobs[[i]]$grobs[[1]]$childrenOrder))
    g$grobs[[i]]$grobs[[1]]$children[[j]]$gp$fill <- fillsr[k]
    k <- k+1
  }
  k <- 1
  for (i in stript) {
    j <- which(grepl('rect', g$grobs[[i]]$grobs[[1]]$childrenOrder))
    g$grobs[[i]]$grobs[[1]]$children[[j]]$gp$fill <- fillst[k]
    k <- k+1
  }
  return(g)
}


plotORs = function(geneids, hits, title){
  Resall=data.frame()
  
  for (gl in names(genelists)){
    tmp=table(signif=geneids %in% hits, targetgene=geneids %in% genelists[[gl]]$ENTREZ_ID)
    resfish=fisher.test(tmp)
    res = c(resfish$estimate, unlist(resfish$conf.int), resfish$p.value)
    Resall = rbind(Resall, res)
  }
  colnames(Resall)=c("OR", "CI95L", "CI95U", "P")
  rownames(Resall)=names(genelists)
  Resall$Beta = log(Resall$OR)
  Resall$SE = (log(Resall$OR)-log(Resall$CI95L))/1.96
  Resall$Padj=p.adjust(Resall$P, method = "bonferroni")
  
  Resall$label=rownames(Resall) 
  Resall$label <- factor(Resall$label, levels=rev(Resall$label))
  
  starpval=convertpvaltostars(Resall$P)
  starpval[Resall$Padj<0.05] ="adj.p**"
  starpval[is.na(Resall$Beta)]="n.a."
  Resall$starpval = starpval
  
  cis = data.frame(Resall$Beta-1.96*Resall$SE, Resall$Beta+1.96*Resall$SE)
  cis[is.na(cis[1]), 1]=-Inf
  cis[is.na(cis[2]), 2]=Inf
  
  Resall$CIlinmin = apply(cis, 1, min)
  Resall$CIlinmax = apply(cis, 1, max)
  
  
  p= ggplot(Resall, aes(x=label, y=Beta, ymin=CIlinmin, ymax=CIlinmax))+
    geom_pointrange(aes(color=starpval)) + geom_hline(yintercept = 0, lty=2) +
    scale_color_viridis_d(limits=c("", "*", "**", "adj.p**"), end = 0.8)+ 
    coord_flip(ylim=c( -2.5, 2.5)) + xlab("Genelist")+ylab(expression(log(OR) %+-% "95%CI"))+theme_bw()+
    guides(color=guide_legend(title="P-value"))+ggtitle(title)
  
  
  return (p)
  
}


```


```{r Figure1, fig.width = 12, fig.height = 15}


brown_ME=ggplot(coldata, aes(gRNA, MEbrown, fill=gRNA)) +
  geom_boxplot()+facet_grid(DIFF~RAPA, scales="fixed", drop = T)+theme_bw()+ylab("Eigengenevalue")+
  scale_fill_manual(values=c(Dark8[2:4]))+ theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))+
  ggtitle("brown")+ theme(plot.title = element_text(hjust = 0.5))


brown_ME=recolor(brown_ME)


module = "brown"  
idx = GOresWGCNA$query==module & grepl("GO:BP", GOresWGCNA$source)
if(!any(idx)){
  p= ggplot() + annotate("text", x = 4, y = 25, size=4, 
                         label = "no significant GO term") + 
    ggtitle(module)+theme_void()+
    theme(plot.title = element_text(hjust = 0.5))
} else {
  p=GOplot(GOresWGCNA[idx, ], 10, Title =module, ylabel = "GO terms")+ theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
}

brown_GO=p

idx = GOresWGCNA$query==module & grepl("TF", GOresWGCNA$source)
if(!any(idx)){
  p= ggplot() + annotate("text", x = 4, y = 25, size=4, 
                         label = "no significant TFs") + 
    ggtitle(module)+theme_void()+
    theme(plot.title = element_text(hjust = 0.5))
} else {
  p=GOplot(GOresWGCNA[idx, ], 10, Title =module, ylabel = "Transcription factors")+ theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
}

brown_TF=p

brown_OR = plotORs(geneids = rownames(metadata), hits = rownames(metadata)[metadata$cluster=="brown"], title = "brown")

black_ME=ggplot(coldata, aes(gRNA, MEblack, fill=gRNA)) +
  geom_boxplot()+facet_grid(DIFF~RAPA, scales="free", drop = T)+theme_bw()+ylab("Eigengenevalue")+
  scale_fill_manual(values=c(Dark8[2:4]))+ theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))+
  ggtitle("black")+ theme(plot.title = element_text(hjust = 0.5))
black_ME = recolor(black_ME)

module = "black"  
idx = GOresWGCNA$query==module & grepl("GO:BP", GOresWGCNA$source)
if(!any(idx)){
  p= ggplot() + annotate("text", x = 4, y = 25, size=4, 
                         label = "no significant GO term") + 
    ggtitle(module)+theme_void()+
    theme(plot.title = element_text(hjust = 0.5))
} else {
  p=GOplot(GOresWGCNA[idx, ], 10, Title =module, ylabel = "GO terms")+ theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
}

black_GO=p

idx = GOresWGCNA$query==module & grepl("TF", GOresWGCNA$source)
if(!any(idx)){
  p= ggplot() + annotate("text", x = 4, y = 25, size=4, 
                         label = "no significant TFs") + 
    ggtitle(module)+theme_void()+
    theme(plot.title = element_text(hjust = 0.5))+ylab("Transcription factors")
} else {
  p=GOplot(GOresWGCNA[idx, ], 10, Title =module, ylabel = "Transcription factors")+ theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
}

black_TF=p
black_OR = plotORs(geneids = rownames(metadata), hits = rownames(metadata)[metadata$cluster=="black"], title = "brown")

X = ggplot() + theme_void()



finplot = ggpubr::ggarrange(
  brown_ME, X, black_ME,
  X,  X,  X, 
  brown_GO,  X, black_GO,
  X,  X,  X, 
  brown_TF,  X, black_TF,
  X,  X,  X, 
  brown_OR,  X, black_OR, 
  labels = c("A", "", "E","", "", "",  "B", "", "F", "", "", "", "C", "", "G","", "", "",  "D", "", "H"), 
  ncol = 3, nrow = 7 ,widths = c(1,0.1,1), heights = c(0.8, 0.1, 1, 0.1, 1, 0.1, 1.2))


pdf(paste0("./docs/Publication_figure1.pdf"), width = 12, height = 15)
finplot
dev.off()

ggsave("./docs/Publication_figure1.svg", finplot, width = 10, height = 13)



```



```{r Figure 2}

# create plots 

prolif_heatmap = geneheatmap(GOIsEntrez = prolif_dex_results$ENTREZ_ID[which(prolif_dex_results$REPLICATED)],
                             exprobj = ddsMat[,ddsMat$DIFF=="noDIFF"], 
                             hm.cluster_rows = T, 
                             hm.scale = "row",  
                             annotation_row_labels = rev(c(15,20, 24)), 
                             annotation_row_names = rev(c("DEPDC5_KO_dr", "Epilepsy", "ASD")), 
                             # index of genelist to plot
                             title = "PROLIF DEX") %>% as.ggplot(hjust = .1)

idx = grepl("GO:BP", prolifGO_results$source)
prolif_GOplot=GOplot(prolifGO_results[idx, ], 10, Title="PROLIF DEX", ylabel = "GO terms")+ 
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

idx = grepl("TF", prolifGO_results$source)
if(!any(idx)){
  prolif_TFplot= ggplot() + annotate("text", x = 4, y = 25, size=4, 
                                     label = "no significant TF") + 
    ggtitle("PROLIF DEX")+theme_void()+
    theme(plot.title = element_text(hjust = 0.5))
} else {
  prolif_TFplot=GOplot(prolifGO_results[idx, ], 10, Title="PROLIF DEX", ylabel = "Transcription Factors")+ 
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
}

prolif_OR = plotORs(geneids = rownames(metadata), 
                    hits = prolif_dex_results$ENTREZ_ID[prolif_dex_results$REPLICATED], 
                    title = "PROLIF DEX")

diff_heatmap = geneheatmap(GOIsEntrez = diff_dex_results$ENTREZ_ID[which(diff_dex_results$REPLICATED)],
                           exprobj = ddsMat[,ddsMat$DIFF=="DIFF"], 
                           hm.cluster_rows = T, 
                           hm.scale = "row", 
                           annotation_row_labels = rev(c(15,20,24)), 
                           annotation_row_names = rev(c("DEPDC5_KO_dr", "Epilepsy", "ASD")),
                           title = "DIFF DEX") %>% as.ggplot(hjust = .1)

idx = grepl("GO:BP", diffGO_results$source)
diff_GOplot=GOplot(diffGO_results[idx, ], 10, Title="DIFF DEX", ylabel = "GO terms")+ 
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

idx = grepl("TF", diffGO_results$source)
diff_TFplot=GOplot(diffGO_results[idx, ], 10, Title="DIFF DEX", ylabel = "Transcription Factors")+ 
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))


diff_OR = plotORs(geneids = rownames(metadata), 
                  hits = diff_dex_results$ENTREZ_ID[diff_dex_results$REPLICATED], 
                  title = "DIFF DEX")


finplot = ggarrange(
  prolif_heatmap,X, diff_heatmap, 
  X,  X,  X, 
  prolif_GOplot,X, diff_GOplot, 
  X,  X,  X, 
  prolif_TFplot,X, diff_TFplot, 
  X,  X,  X, 
  prolif_OR, X, diff_OR,
  labels = c("A", "", "E","", "", "",  "B", "", "F", "", "", "", "C", "", "G","", "", "",  "D", "", "H"), 
  ncol = 3, nrow = 7, widths = c(1, 0.1, 1), heights = c(1.5, 0.1, 1, 0.1, 1, 0.1, 1.3))


pdf(paste0("./docs/Publication_figure2.pdf"), width = 12, height = 15)
finplot
dev.off()

ggsave("./docs/Publication_figure2.svg", finplot, width = 10, height = 13)



```

```{r Figure 3, fig.width= 10, fig.height=15}
idxnoresc = grepl("GO:BP", Rescue_PROLIF_GO_results$source) & grepl("RAPA_noRESC", Rescue_PROLIF_GO_results$query) 
idxresc = grepl("GO:BP", Rescue_PROLIF_GO_results$source) & grepl("RAPA_RESC", Rescue_PROLIF_GO_results$query) 


PROLIF_GO_BP_RESC_GOplot=GOplot(Rescue_PROLIF_GO_results[idxresc, ], 10, Title="PROLIF RAPA RESC", ylabel = "GO terms")+ 
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))



PROLIF_GO_BP_noRESC_GOplot=GOplot(Rescue_PROLIF_GO_results[idxnoresc, ], 10, Title="PROLIF RAPA noRESC", ylabel = "GO terms")+ 
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))



idxnoresc = grepl("TF", Rescue_PROLIF_GO_results$source) & grepl("RAPA_noRESC", Rescue_PROLIF_GO_results$query) 
idxresc = grepl("TF", Rescue_PROLIF_GO_results$source) & grepl("RAPA_RESC", Rescue_PROLIF_GO_results$query) 

if(!any(idxresc)){
  PROLIF_GO_TF_RESC_GOplot= ggplot() + annotate("text", x = 4, y = 25, size=4, 
                                                label = "no significant TF") + 
    ggtitle("PROLIF RAPA RESC")+theme_void()+
    theme(plot.title = element_text(hjust = 0.5))
} else {
  PROLIF_GO_TF_RESC_GOplot=GOplot(Rescue_PROLIF_GO_results[idxresc, ], 10, Title="PROLIF RAPA RESC", ylabel = "Transcription Factors")+ 
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
}


if(!any(idxnoresc)){
  PROLIF_GO_TF_noRESC_GOplot= ggplot() + annotate("text", x = 4, y = 25, size=4, 
                                                  label = "no significant TF") + 
    ggtitle("PROLIF RAPA noRESC")+theme_void()+
    theme(plot.title = element_text(hjust = 0.5))
} else {
  PROLIF_GO_TF_noRESC_GOplot=GOplot(Rescue_PROLIF_GO_results[idxnoresc, ], 10, Title="PROLIF RAPA noRESC", ylabel = "Transcription Factors")+ 
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
}

PROLIF_GO_BP_RESC_ORplot =  plotORs(geneids = rownames(metadata), 
                                    hits = Rescue_Genelist[["PROLIF"]][["noDIFFRAPA_RESC"]], 
                                    title = "PROLIF RAPA RESC")
PROLIF_GO_BP_noRESC_ORplot =  plotORs(geneids = rownames(metadata), 
                                      hits = Rescue_Genelist[["PROLIF"]][["noDIFFRAPA_noRESC"]], 
                                      title = "PROLIF RAPA noRESC")
```

```{r Figure 3 DIFF, fig.width= 10, fig.height=15}
##### DIFF

idxnoresc = grepl("GO:BP", Rescue_DIFF_GO_results$source) & grepl("RAPA_noRESC", Rescue_DIFF_GO_results$query) 
idxresc = grepl("GO:BP", Rescue_DIFF_GO_results$source) & grepl("RAPA_RESC", Rescue_DIFF_GO_results$query) 


DIFF_GO_BP_RESC_GOplot=GOplot(Rescue_DIFF_GO_results[idxresc, ], 10, Title="DIFF RAPA RESC", ylabel = "GO terms")+ 
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

DIFF_GO_BP_noRESC_GOplot=GOplot(Rescue_DIFF_GO_results[idxnoresc, ], 10, Title="DIFF RAPA noRESC", ylabel = "GO terms")+ 
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))


idxnoresc = grepl("TF", Rescue_DIFF_GO_results$source) & grepl("RAPA_noRESC", Rescue_DIFF_GO_results$query) 
idxresc = grepl("TF", Rescue_DIFF_GO_results$source) & grepl("RAPA_RESC", Rescue_DIFF_GO_results$query) 

if(!any(idxresc)){
  DIFF_GO_TF_RESC_GOplot= ggplot() + annotate("text", x = 4, y = 25, size=4, 
                                              label = "no significant TF") + 
    ggtitle("DIFF RAPA RESC")+theme_void()+
    theme(plot.title = element_text(hjust = 0.5))
} else {
  DIFF_GO_TF_RESC_GOplot=GOplot(Rescue_DIFF_GO_results[idxresc, ], 10, Title="DIFF RAPA RESC", ylabel = "Transcription Factors")+ 
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
}


if(!any(idxnoresc)){
  DIFF_GO_TF_noRESC_GOplot= ggplot() + annotate("text", x = 4, y = 25, size=4, 
                                                label = "no significant TF") + 
    ggtitle("DIFF RAPA noRESC")+theme_void()+
    theme(plot.title = element_text(hjust = 0.5))
} else {
  DIFF_GO_TF_noRESC_GOplot=GOplot(Rescue_DIFF_GO_results[idxnoresc, ], 10, Title="DIFF RAPA noRESC", ylabel = "Transcription Factors")+ 
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
}

DIFF_GO_BP_RESC_ORplot =  plotORs(geneids = rownames(metadata), 
                                  hits = Rescue_Genelist[["DIFF"]][["DIFFRAPA_RESC"]], 
                                  title = "DIFF RAPA RESC")
DIFF_GO_BP_noRESC_ORplot =  plotORs(geneids = rownames(metadata), 
                                    hits = Rescue_Genelist[["DIFF"]][["DIFFRAPA_noRESC"]], 
                                    title = "DIFF RAPA noRESC")

X = ggplot() + theme_void()

finplot = ggpubr::ggarrange(
  PROLIF_GO_BP_RESC_GOplot, X, DIFF_GO_BP_RESC_GOplot,
  X,  X,  X, 
  PROLIF_GO_BP_noRESC_GOplot,  X, DIFF_GO_BP_noRESC_GOplot,
  X,  X,  X, 
  PROLIF_GO_TF_RESC_GOplot,  X, DIFF_GO_TF_RESC_GOplot,
  X,  X,  X, 
  PROLIF_GO_TF_noRESC_GOplot,  X, DIFF_GO_TF_noRESC_GOplot, 
  X,  X,  X, 
  PROLIF_GO_BP_RESC_ORplot, X, DIFF_GO_BP_RESC_ORplot, 
  X,  X,  X, 
  PROLIF_GO_BP_noRESC_ORplot, X, DIFF_GO_BP_noRESC_ORplot,
  
  labels = c("A", "", "G",
             "", "", "",  
             "B", "", "H", 
             "", "", "", 
             "C", "", "I",
             "", "", "",  
             "D", "", "J",
             "", "", "",  
             "E", "", "K",
             "", "", "",  
             "F", "", "L"), 
  ncol = 3, nrow = 11 ,widths = c(1,0.1,1), 
  heights = c(0.7, 0.1, 0.9, 0.1, 0.2, 0.1, 0.7, 0.1,1.2,0.1,1.2))


pdf(paste0("./docs/Publication_figure3.pdf"), width = 12, height = 15)
finplot
dev.off()

ggsave("./docs/Publication_figure3.svg", finplot, width = 10, height = 13)


```



```{r Supplementary Figure 3, fig.width=10, fig.height=10}
reads= counts(ddsMat)

testset=reads[sample(1:nrow(reads), 1000),]

sums=data.frame(log2_total_reads=log2(colSums(reads)), label=colnames(reads))
sums$Sample = coldata[sums$label,"label"]

SumReadplot = ggplot(data=sums,aes(x=label, y=log2_total_reads, fill=Sample))+
  geom_bar(stat = "identity")+theme_classic()+
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank())+
  scale_fill_brewer(type = "qual", palette ="Set3")


longdata = reshape2::melt(reads)
colnames(longdata) = c("ENTREZ_ID", "Sample", "Counts")
longdata$label = coldata[longdata$Sample, "label"]

Readsdistplot = ggplot(data=longdata, aes(y=log2(Counts+1), x=Sample, fill=label))+
  geom_boxplot(notch = T)+theme_classic()+
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank())+
  scale_fill_brewer(type = "qual", palette ="Set3")


reads= counts(ddsMat, normalized=T)
vars=apply(log2(reads+1), 1, var, na.rm=T)
topvars = log2(reads[which(vars %in% sort(vars, decreasing=T)[1:2000]),]+1)

distance = dist(t(topvars))

sampleDistMatrix <- as.matrix(distance)

#colors for plotting heatmap
colors <- viridis::viridis(255)

SampleInfo=coldata
cellcol = Dark8[1:nlevels(SampleInfo$CellLine)]
names(cellcol) = levels(SampleInfo$CellLine)

gRNAcol = Dark8[c(1:nlevels(SampleInfo$gRNA))+nlevels(SampleInfo$CellLine)]
names(gRNAcol) = levels(SampleInfo$gRNA)

diffcol = brewer.pal(3,"Set1")[1:nlevels(SampleInfo$DIFF)]
names(diffcol) = levels(SampleInfo$DIFF)

rapacol = brewer.pal(3,"Set2")[1:nlevels(SampleInfo$RAPA)]
names(rapacol) = levels(SampleInfo$RAPA)

ann_colors = list(
  DIFF = diffcol, 
  RAPA = rapacol,
  gRNA = gRNAcol
  #,CellLine=cellcol
)

labels = SampleInfo[,c("gRNA","DIFF", "RAPA")] %>%  
  mutate_all(as.character) %>% as.data.frame()

rownames(labels)=SampleInfo$label_rep
colnames(sampleDistMatrix) = SampleInfo[rownames(sampleDistMatrix), "label_rep"]
rownames(sampleDistMatrix)=colnames(sampleDistMatrix)

heatmapplot = pheatmap(sampleDistMatrix,
                       clustering_distance_rows = distance,
                       clustering_distance_cols = distance,
                       clustering_method = "ward.D2",
                       scale ="row",
                       border_color = NA, 
                       annotation_row = labels,
                       annotation_col = labels,
                       annotation_colors = ann_colors,
                       col = colors, show_colnames = F, show_rownames = F,
                       main = "Distances normalized log2 counts")



finplot = ggarrange(ggarrange(SumReadplot, Readsdistplot,
                              labels=c("A", "B")), 
                    as.grob(heatmapplot), ncol = 1, labels = c("", "C"), 
                    heights = c(1,2))


pdf(paste0("./docs/Publication_Supplfigure3.pdf"), width = 12, height = 15)
finplot
dev.off()

ggsave("./docs/Publication_Supplfigure3.svg", finplot, width = 10, height = 13)


```

```{r}

# supplTab 5
resfile = grep("GO", list.files("C:/Users/andreas_chiocchetti/OneDrive/Documents/Frankfurt Uni/Jones/Manuscript_Jones/DEPDC5_D62_Analysis/output"), value = T)

names(resfile) = gsub(".xlsx|GOres|D62|_", "", resfile)
names(resfile) = gsub("noDIFF", "PROLIF", names(resfile))


datset= list()
for(n in resfile){
  datset[[n]] = openxlsx::read.xlsx(paste0("./output/", n))
}

names(datset) = names(resfile)

write.xlsx(datset, file = "./output/Table_S5_GOres_combined.xlsx")



```

```{r}

# supplTab 3
resfile = grep("Restab", list.files("C:/Users/andreas_chiocchetti/OneDrive/Documents/Frankfurt Uni/Jones/Manuscript_Jones/DEPDC5_D62_Analysis/output"), value = T)

names(resfile) = gsub(".xlsx|Restab|D62|_", "", resfile)
names(resfile) = gsub("noDIFF", "PROLIF", names(resfile))

resfile = grep("NTCnoRAPA", resfile, value = T)


system("rm ./output/Table_S3_DEX_results_combined.xlsx")


datset= list()
for(n in resfile){
  datset[[n]] = openxlsx::read.xlsx(paste0("./output/", n))
}

names(datset) = names(resfile)
write.xlsx(datset, file = "./output/Table_S3_DEXres_combined.xlsx")

```
